backend/requirements.txt
# ========== Core Framework ==========
Django==4.2.23
djangorestframework==3.16.1
djangorestframework-simplejwt
django-allauth==65.11.0
dj-rest-auth==7.0.1
django-filter==25.1
django-cors-headers==4.7.0
django-redis==6.0.0
asgiref==3.9.1
sqlparse==0.5.3
python-decouple==3.8
python-dotenv==1.1.1
psycopg2-binary==2.9.10
redis==6.4.0

# ========== Authentication & JWT ==========
PyJWT==2.10.1
cryptography==45.0.6

# ========== AI & Google API Integration ==========
google-generativeai==0.8.5
google-genai==1.31.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
proto-plus==1.26.1
protobuf==5.29.5
httplib2==0.22.0
uritemplate==4.2.0

# ========== Utilities ==========
requests==2.32.5
urllib3==2.5.0
idna==3.10
charset-normalizer==3.4.3
tqdm==4.67.1
pillow==11.3.0
typing_extensions==4.14.1
typing-inspection==0.4.1
cachetools==5.5.2
anyio==4.10.0
httpcore==1.0.9
httpx==0.28.1
sniffio==1.3.1
websockets==15.0.1
tenacity==9.1.2
pyparsing==3.2.3
cffi==1.17.1
pycparser==2.22
rsa==4.9.1
exceptiongroup==1.3.0
annotated-types==0.7.0
pydantic==2.11.7
pydantic_core==2.33.2

# ========== Testing & Coverage ==========
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-html==4.1.1
coverage==7.6.8
Faker==30.6.0
factory-boy==3.3.1

# ========== Optional: Async / Timeout Tools ==========
async-timeout==5.0.1
